Fix slow running WebSocket tests
This reduces the runtime for the TCK by more than 10 minutes on my dev
box. The tests were using a CountDownLatch incorrectly to try and wait
for messages to arrive before continuing. The result was, effectively,
lots of unnecessary 5s sleeps during tests that amounted to more than 10
minutes of delay.
This commit fixes those uses of CountDownLatch so the waits only wait
for the messages that are expected and adds a utility class that waits
for sessions to close.
A happy side-effect is that a couple a timeout tests that failed
frequently now pass consistently.

Signed-off-by: Mark Thomas <markt@apache.org>