Correct the response code being looked for.
There is nothing in the Servlet specification that suggests just adding
"/j_security_check" to the end of a URI should trigger authentication.
Depending on the servlet mapping (exact or path mapping) the request
will either fail with 404 or work with "/j_security_check" available as
a path parameter. In this case an exact mapping is used so 404 is the
correct response.

Signed-off-by: Mark Thomas <markt@apache.org>