Restore copyright for java files not updated this year
Done by comparing the file contents from the `9.0.0` tag against
current contents. For any files that had the exact same contents, sans
copyright, the old file was restore as-is.

This is so we can see what has actually changed between the EE 9.0 and
current 9.1 branches.  480 files were updated in copyright only.  We
should either update all files or no files.

Signed-off-by: David Blevins <david.blevins@gmail.com>